+++ tests/test_concat.py
@@ -34,6 +34,7 @@ class TestConcat(TestCase):
         self,
         data_dir,
         output_name,
+        concat_method: str = "xarray-concat",
         record_dim_name: str = "mirror_step",
         concat_kwargs: dict | None = None,
     ):
@@ -55,6 +56,7 @@ class TestConcat(TestCase):
             output_file=output_path,
             write_tmp_flat_concatenated=True,
             keep_tmp_files=True,
+            concat_method=concat_method,
             concat_dim=record_dim_name,
             concat_kwargs=concat_kwargs,
         )
@@ -93,16 +95,24 @@ class TestConcat(TestCase):
         assert length_sum == len(merged_dataset.variables[record_dim_name])
 
     def test_tempo_no2_concat_with_stitchee(self):
-        self.run_verification_with_stitchee("tempo/no2", "tempo_no2_bee_concatenated.nc")
+        self.run_verification_with_stitchee(
+            "tempo/no2", "tempo_no2_bee_concatenated.nc", concat_method="xarray-concat"
+        )
 
     def test_tempo_hcho_concat_with_stitchee(self):
-        self.run_verification_with_stitchee("tempo/hcho", "tempo_hcho_bee_concatenated.nc")
+        self.run_verification_with_stitchee(
+            "tempo/hcho", "tempo_hcho_bee_concatenated.nc", concat_method="xarray-concat"
+        )
 
     def test_tempo_cld04_concat_with_stitchee(self):
-        self.run_verification_with_stitchee("tempo/cld04", "tempo_cld04_bee_concatenated.nc")
+        self.run_verification_with_stitchee(
+            "tempo/cld04", "tempo_cld04_bee_concatenated.nc", concat_method="xarray-concat"
+        )
 
     def test_tempo_o3prof_concat_with_stitchee(self):
-        self.run_verification_with_stitchee("tempo/o3prof", "tempo_o3prof_bee_concatenated.nc")
+        self.run_verification_with_stitchee(
+            "tempo/o3prof", "tempo_o3prof_bee_concatenated.nc", concat_method="xarray-concat"
+        )
 
     # def test_icesat_concat_with_stitchee(self):
     #     self.run_verification_with_stitchee('icesat', 'icesat_concat_with_stitchee.nc')
@@ -111,15 +121,18 @@ class TestConcat(TestCase):
         self.run_verification_with_stitchee(
             "ceres-subsetter-output",
             "ceres_bee_concatenated.nc",
+            concat_method="xarray-combine",
             record_dim_name="time",
-            concat_kwargs={"compat": "override"},
+            concat_kwargs={"compat": "override", "combine_attrs": "override"},
         )
 
     def test_ceres_flash_concat_with_stitchee(self):
         self.run_verification_with_stitchee(
             "ceres_flash-subsetter-output",
             "ceres_flash_bee_concatenated.nc",
+            concat_method="xarray-combine",
             record_dim_name="time",
+            concat_kwargs={"compat": "override", "combine_attrs": "override"},
         )
 
     # def test_ceres_flash_concat_with_stitchee(self):